aosdjh2

+++ Software/FlexPMS/KarContainer.cpp
@@ -59,9 +59,10 @@ void KarContainer::reload() {
         kar->id = res->getInt("id");
         kar->name = res->getString("name");
         kar->address = res->getInt("address");
-        kar->ph = res->getInt("ph");
+        kar->ph = res->getDouble("ph");
         kar->volumen = res->getInt("volumen");
         kar->humidity = res->getInt("humidity");
+        
         kar->mwstatus = res->getBoolean("mwstatus");
         kar->ivalvestatus = res->getBoolean("ivalvestatus");
         kar->ovalvestatus = res->getBoolean("ovalvestatus");
